+++ src/main/kotlin/com/github/jensim/megamanipulator/actions/vcs/bitbucketserver/BitBucketTypes.kt
@@ -2,6 +2,7 @@ package com.github.jensim.megamanipulator.actions.vcs.bitbucketserver
 
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.Transient
+import kotlinx.serialization.json.JsonElement
 
 @Serializable
 data class BitBucketMessage(
@@ -11,12 +12,12 @@ data class BitBucketMessage(
 )
 
 @Serializable
-data class BitBucketPage<T>(
+data class BitBucketPage(
     val start: Int? = null,
     val size: Int? = null,
     val limit: Int? = null,
     val isLastPage: Boolean? = null,
-    val values: List<T>? = null,
+    val values: List<JsonElement>? = null,
     val message: String? = null,
 )